+++ tanjun/abc.py
@@ -498,6 +498,15 @@ class Context(abc.ABC):
             A singular attachment to edit the response with.
         attachments : hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]]
             A sequence of attachments to edit the response with.
+        component : hikari.undefined.UndefinedNoneOr[hikari.api.special_endpoints.ComponentBuilder]
+            If provided, builder object of the component to set for this message.
+            This component will replace any previously set components and passing
+            `builtins.None` will remove all components.
+        components : hikari.undefined.UndefinedNoneOr[typing.Sequence[hikari.api.special_endpoints.ComponentBuilder]]
+            If provided, a sequence of the component builder objects set for
+            this message. These components will replace any previously set
+            components and passing `builtins.None` or an empty sequence will
+            remove all components.
         embed : hikari.UndefinedOr[hikari.Embed]
             An embed to replace the response with.
         embeds : hikari.UndefinedOr[collections.Sequence[hikari.Embed]]
@@ -635,6 +644,15 @@ class Context(abc.ABC):
             A singular attachment to edit the response with.
         attachments : hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]]
             A sequence of attachments to edit the response with.
+        component : hikari.undefined.UndefinedNoneOr[hikari.api.special_endpoints.ComponentBuilder]
+            If provided, builder object of the component to set for this message.
+            This component will replace any previously set components and passing
+            `builtins.None` will remove all components.
+        components : hikari.undefined.UndefinedNoneOr[typing.Sequence[hikari.api.special_endpoints.ComponentBuilder]]
+            If provided, a sequence of the component builder objects set for
+            this message. These components will replace any previously set
+            components and passing `builtins.None` or an empty sequence will
+            remove all components.
         embed : hikari.UndefinedOr[hikari.Embed]
             An embed to replace the response with.
         embeds : hikari.UndefinedOr[collections.Sequence[hikari.Embed]]
@@ -830,6 +848,11 @@ class Context(abc.ABC):
         ----------------
         ensure_result : bool
             If provided and set to `True`, It may perform an extra API call
+        component : hikari.undefined.UndefinedOr[hikari.api.special_endpoints.ComponentBuilder]
+            If provided, builder object of the component to include in this message.
+        components : hikari.undefined.UndefinedOr[typing.Sequence[hikari.api.special_endpoints.ComponentBuilder]]
+            If provided, a sequence of the component builder objects to include
+            in this message.
         embed : hikari.UndefinedOr[hikari.Embed]
             An embed to respond with.
         embeds : hikari.UndefinedOr[collections.Sequence[hikari.Embed]]
@@ -1038,6 +1061,11 @@ class MessageContext(Context, abc.ABC):
             A singular attachment to respond with.
         attachments : hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]]
             A sequence of attachments to respond with.
+        component : hikari.undefined.UndefinedOr[hikari.api.special_endpoints.ComponentBuilder]
+            If provided, builder object of the component to include in this message.
+        components : hikari.undefined.UndefinedOr[typing.Sequence[hikari.api.special_endpoints.ComponentBuilder]]
+            If provided, a sequence of the component builder objects to include
+            in this message.
         embed : hikari.UndefinedOr[hikari.Embed]
             An embed to respond with.
         embeds : hikari.UndefinedOr[collections.Sequence[hikari.Embed]]